重构为int64！

+++ common/safe_map.go
@@ -7,17 +7,17 @@ import (
 // SafeMap 安全的map，支持并发读写
 type SafeMap struct {
 	sync.RWMutex
-	m    map[int]interface{}
-	keys []int
+	m    map[int64]interface{}
+	keys []int64
 }
 
 // NewSafeMap 创建一个安全的map
 func NewSafeMap() *SafeMap {
-	return &SafeMap{m: make(map[int]interface{})}
+	return &SafeMap{m: make(map[int64]interface{})}
 }
 
-// 设置值
-func (sm *SafeMap) Set(key int, value interface{}) {
+// Set 设置值
+func (sm *SafeMap) Set(key int64, value interface{}) {
 	sm.Lock()
 	defer sm.Unlock()
 	if _, ok := sm.m[key]; !ok {
@@ -26,22 +26,22 @@ func (sm *SafeMap) Set(key int, value interface{}) {
 	sm.m[key] = value
 }
 
-// 获取值
-func (sm *SafeMap) Get(key int) (interface{}, bool) {
+// Get 获取值
+func (sm *SafeMap) Get(key int64) (interface{}, bool) {
 	sm.RLock()
 	defer sm.RUnlock()
 	value, ok := sm.m[key]
 	return value, ok
 }
 
-// 获取所有键
-func (sm *SafeMap) Keys() []int {
+// Keys 获取所有键
+func (sm *SafeMap) Keys() []int64 {
 	sm.RLock()
 	defer sm.RUnlock()
 	return sm.keys
 }
 
-// 获取所有值
+// Values 获取所有值
 func (sm *SafeMap) Values() []interface{} {
 	sm.RLock()
 	defer sm.RUnlock()
